refactor(street-name-checker): use the shared DOM builders
Both features now build their markup from src/ui/dom.ts, and the two per-feature
dom.ts files are gone.

The four settings subsections take a Waze icon name here rather than in the icon
commit: buildSubsection now requires a name where it used to take an emoji, so
there is no intermediate state to pass through.
